+++ src/Plugin.php
@@ -8,6 +8,7 @@ use benf\neo\controllers\Input as InputController;
 use benf\neo\elements\Block;
 use benf\neo\fieldlayoutelements\ChildBlocksUiElement;
 use benf\neo\gql\interfaces\elements\Block as NeoGqlInterface;
+use benf\neo\integrations\feedme\Field as FeedMeField;
 use benf\neo\models\Settings;
 use benf\neo\services\Blocks as BlocksService;
 use benf\neo\services\BlockTypes as BlockTypesService;
@@ -27,6 +28,8 @@ use craft\events\RegisterComponentTypesEvent;
 use craft\events\RegisterConditionRuleTypesEvent;
 use craft\events\RegisterGqlTypesEvent;
 use craft\events\RegisterUserPermissionsEvent;
+use craft\feedme\events\RegisterFeedMeFieldsEvent;
+use craft\feedme\services\Fields as FeedMeFields;
 use craft\gatsbyhelper\events\RegisterIgnoredTypesEvent;
 use craft\gatsbyhelper\services\Deltas;
 use craft\models\FieldLayout;
@@ -105,6 +108,7 @@ class Plugin extends BasePlugin
         $this->_registerResaveBlocksCommand();
         $this->_registerPermissions();
         $this->_registerGatsbyHelper();
+        $this->_registerFeedMeSupport();
         $this->_registerConditionFieldRuleRemoval();
     }
 
@@ -268,6 +272,17 @@ class Plugin extends BasePlugin
                 $event->types[] = Block::class;
             });
         }
+    }
+
+    private function _registerFeedMeSupport(): void
+    {
+        Event::on(
+            FeedMeFields::class,
+            FeedMeFields::EVENT_REGISTER_FEED_ME_FIELDS,
+            function(RegisterFeedMeFieldsEvent $e) {
+                $e->fields[] = FeedMeField::class;
+            }
+        );
     }
 
     /**